code clean-up
